products_service/requirements.txt
flask
flask-httpauth
mysql-connector-python
pytest
requests
gunicorn